fix(create): stop NetFloorArea exceeding GrossFloorArea under --boundary outer
ifc-lite generate-spaces --boundary outer bakes the emitted IfcSpace at the
outward-offset (gross) wall face, but the orchestrator only passed
grossFloorArea (the centreline measure) to addSpaceToStore. addSpaceToStore
falls back to the OuterCurve polygon's own area for NetFloorArea when a
caller omits it — an assumption that only holds for the default
--boundary inner, where OuterCurve genuinely is the inner face. Under
'outer' the reported NetFloorArea came out larger than GrossFloorArea,
which Qto_SpaceBaseQuantities never allows.

generateSpacesFromWalls now always derives netFloorArea from the inner
(room-side) inset, independent of boundaryMode. The geometry offset
helpers moved unchanged to a new sibling module, room-footprint-offset.ts,
to keep generate-spaces.ts under the module-size budget; the package's
public exports are unchanged.
