+++ packages/create/src/in-store/space.ts
@@ -81,7 +81,10 @@ export interface SpacePolygonParams {
   /** Bounding elements → one IfcRelSpaceBoundary each. */
   boundaries?: SpaceBoundaryInput[];
   /** Net (inner-face) floor area in m² for Qto_SpaceBaseQuantities; defaults
-   *  to the gross/centreline area when omitted. */
+   *  to the OuterCurve polygon's own area when omitted — pass this explicitly
+   *  whenever OuterCurve is not itself the inner (room-side) face, e.g. a
+   *  centreline or outer-face boundary, otherwise NetFloorArea can come out
+   *  larger than GrossFloorArea. */
   netFloorArea?: number;
   /** Gross (centreline) floor area in m² for GrossFloorArea + GrossVolume;
    *  defaults to the OuterCurve area when omitted. */
@@ -195,7 +198,11 @@ export function addSpaceToStore(
   // Qto_SpaceBaseQuantities — attached via the property view (createQuantitySet)
   // rather than as raw IfcElementQuantity entities, so they surface in the
   // properties panel (getQuantitiesForEntity) AND export, from one source.
-  // `area` is the OuterCurve (net when generated from walls) footprint;
+  // `area` is the OuterCurve polygon's own footprint — the fallback used only
+  // when a caller omits `netFloorArea` / `grossFloorArea`. A caller whose
+  // OuterCurve isn't the inner (room-side) face — e.g. a centreline or
+  // outer-face boundary — must pass `netFloorArea` explicitly, else
+  // NetFloorArea comes out equal to or larger than GrossFloorArea.
   // GrossFloorArea/GrossVolume take the supplied centreline measure.
   const area = polygon ? polygonArea(params.OuterCurve) : params.Width * params.Depth;
   const perimeter = polygon
